+++ src/frontend/apps/impress/src/features/docs/doc-management/api/index.ts
@@ -1,3 +1,4 @@
+export * from './collaborationRequests';
 export * from './useCreateDoc';
 export * from './useDeleteFavoriteDoc';
 export * from './useDoc';